+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/BibleReadingRepository.kt
@@ -16,9 +16,11 @@
 
 package me.xizzhu.android.joshua.core.repository
 
+import android.os.Bundle
 import androidx.collection.LruCache
 import me.xizzhu.android.joshua.core.Verse
 import me.xizzhu.android.joshua.core.VerseIndex
+import me.xizzhu.android.joshua.core.analytics.Analytics
 import me.xizzhu.android.joshua.core.repository.local.LocalReadingStorage
 
 class BibleReadingRepository(private val localReadingStorage: LocalReadingStorage) {
@@ -96,6 +98,12 @@ class BibleReadingRepository(private val localReadingStorage: LocalReadingStorag
     suspend fun readVerseWithParallel(translationShortName: String, verseIndex: VerseIndex): Verse =
             localReadingStorage.readVerseWithParallel(translationShortName, verseIndex)
 
-    suspend fun search(translationShortName: String, query: String): List<Verse> =
-            localReadingStorage.search(translationShortName, readBookNames(translationShortName), query)
+    suspend fun search(translationShortName: String, query: String): List<Verse> {
+        Analytics.track(Analytics.EVENT_SEARCH, Bundle().apply {
+            putString(Analytics.PARAM_ITEM_ID, translationShortName)
+            putString(Analytics.PARAM_SEARCH_TERM, query)
+        })
+
+        return localReadingStorage.search(translationShortName, readBookNames(translationShortName), query)
+    }
 }